requirements.txt
moderngl
pygame
pywavefront
pyglm
numpy